incident_manager/pyproject.toml
[project]
name = "incident-manager"
version = "0.1.0"
description = ""
authors = [
    {name = "Soliton"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "llama-index-vector-stores-milvus (>=0.8.3,<0.9.0)",
    "llama-index-embeddings-openai (>=0.3.1,<0.4.0)"
]
packages = [
    {include = "incident_manager", from = "./"},
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.test.dependencies]
ruff = "^0.11.10"

